Refactor authentication screen component
Also update notification token in firebase after login
